pyproject.toml
[project]
name = "plant-identify"
version = "0.1.0"
description = "identify plants Diameter at Breast Height (DBH) from images using depth estimation"
readme = "README.md"
requires-python = "==3.10.11"
dependencies = [
    "matplotlib>=3.9.4",
    "pandas>=2.3.0",
    "pillow>=12.1.0",
    "plotly>=6.5.1",
    "pycocotools>=2.0.11",
    "python-dotenv>=1.2.1",
    "qdrant-client",
    "sam3",
    "depth-anything-3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "sentencepiece>=0.2.1",
    "triton>=3.6.0; sys_platform == 'linux'",
    "triton-windows>=3.5.1.post24; sys_platform == 'win32'",
    "opencv-python>=4.11.0.86",
    "transformers>=4.57.5",
    "timm>=1.0.24",
    "pydantic>=2.12.5",
    "decord>=0.6.0",
    "einops>=0.8.1",
    "wandb>=0.24.2",
    "torchmetrics>=1.8.2",
    "deeplake>=3.9.50",
    "huggingface-hub>=1.4.1",
    "fastapi>=0.128.3",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.22",
    "pyngrok>=7.5.0",
    "ngrok>=1.4.0",
]

[dependency-groups]
dev = ["ruff>=0.8.2", "nbqa>=1.9.1"]

notebook = ["ipykernel>=6.29.5", "ipywidgets>=8.1.7", "nbformat>=4.2.0"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
torch = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
sam3 = { git = "https://github.com/facebookresearch/sam3.git" }

depth-anything-3 = { git = "https://github.com/ByteDance-Seed/Depth-Anything-3.git" }

[tool.uv]
required-environments = [
    "sys_platform == 'win32' and platform_machine == 'AMD64'",
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402", "B905"] # Allow imports not at top of file in notebooks

[tool.nbqa]
addopts = "--ruff"
